Refactor server code and update model scripts
We have refactored the NoGuiServer.java to modularize, imporove readability and implementation of certain methods. This has made the functions of the code clearer and more maintainable. We have also rearranged the order of methods for cohesive flow. We have added the BaseModel.java, PictureModel.java and updated MessageModel.java for better data handling in the future. Multiple try-except blocks were added to ensure the code behaves correctly under unexpected inputs or in unanticipated situations.