Add table existence check, refactor WebSocket's onOpen and database interaction methods
NoGuiServer.java is enhanced with a provision to check whether a particular table exists in the database before initiating it, which improves database interaction efficiency. Moreover, significant refactoring is performed to compartmentalize functionalities into separate methods. This simplifies the method onOpen in the WebSocket server and breaks down the database interaction logic into smaller, more manageable methods reducing complexity and enhancing maintainability. In addition, the use of Executors to handle separate tasks in own threads improves the performance of the WebSocket's onOpen method as it enables concurrent processing. Changes are also made to use prepared statements which provides SQL Injection protection.