Update codebase and README for new dependencies and DB SSL
This commit hardcodes 'false' for the DB_SSL property in NoGuiServer.java as the SSL feature is yet to be utilised. Also, dependencies in the pom.xml are updated to the newer versions (Java, Maven, Java-WebSocket), and jackson-core is removed as it's merged with other Jackson dependencies. README.md is updated accordingly to reflect these changes in the project's setup. Additionally, instructions for running the project and database are elaborated for clarity. Environment variables required for the application have also been specified in the README.