Add new interface DatabaseConnectionController
Created a new interface, DatabaseConnectionController with methods for communicating with a database. It includes methods for checking if a table exists, initializing the database, setting database settings, retrieving all data, replacing and saving data, retrieving last record from the database, saving messages and images associated with messages, retrieving updated entries. This interface provides a clear contract of communication with the database, improving modularization and allowing possible future extension or variation in the way database communication is handled.