Improve retrieval of environmental variables in DatabaseConnectionController
A new method retrieveEnvironmentVariables() was introduced in the DatabaseConnectionController class. This is to enable retrieval of specific environment variable values easily and in an error-controlled manner. The returned result is wrapped into an optional object to mitigate possible null pointer exceptions due to unavailability of certain environment variables. This system-wide change will make handling of environment-specific data more robust.