Refactored DB env vars handling by introducing a helper function
A helper function, retrieveEnvironmentVariables(), has been added to the DatabaseConnectionController class to simplify and standardize the process of retrieving environment variable values. The method wraps the returned result in an Optional object, mitigating potential null pointer exceptions in case specific environment variables are not set. The goal of this change is to enhance robustness and error handling in the management of environment-specific data.