Remove version from jar final name
The pom.xml was updated to change `finalName` from `teamsocket-${project.version}` to just `teamsocket`. Also, the README.md was adjusted to match this change in instruction for running the .jar file. The change simplifies the build process and usage instructions by removing the need to remember or look up the current project version. The project versioning can still be tracked using git tags.