Add LinkModel handling and refactor BaseModel
Enhanced handling of messages by adding support for LinkModel instances in the MessagesController. This makes the system able to process link messages and store them into the database.

Refactored the BaseModel class to use QuoteModel in place of quotedMessage variables directly. This allows a more systematic approach of storing and processing quoted messages.

In detail:
- The MessagesController now checks and processes LinkModel instances, saving them to the database,
- Introduced @JsonIdentityInfo in BaseModel for managing unique models.
- Removed the old quoted message variables from BaseModel, and replaced them with a more flexible QuoteModel,
- Removed the overriding toString function of BaseModel,
- Updated README and pom.xml for versioning (from v1.0.3 to v1.0.4),
- Introduced a new record class: QuoteModel.